chore: cleanup in Trees.scala
remove unused build targets from constructor and refactor common logic in tests
