+++ tests/unit/src/main/scala/tests/BaseClassFinderSuite.scala
@@ -1,23 +1,13 @@
 package tests
 
 import scala.meta.internal.metals.Buffers
-import scala.meta.internal.metals.BuildTargets
-import scala.meta.internal.metals.ScalaVersionSelector
-import scala.meta.internal.metals.UserConfiguration
 import scala.meta.internal.parsing.ClassFinder
-import scala.meta.internal.parsing.Trees
 
 import munit.FunSuite
 
 abstract class BaseClassFinderSuite extends FunSuite {
   def init(scalaVersion: String): (Buffers, ClassFinder) = {
-    val buffers = Buffers()
-    val buildTargets = new BuildTargets()
-    val selector = new ScalaVersionSelector(
-      () => UserConfiguration(fallbackScalaVersion = Some(scalaVersion)),
-      buildTargets
-    )
-    val trees = new Trees(buildTargets, buffers, selector)
+    val (buffers, trees) = TreeUtils.getTrees(scalaVersion)
     val classFinder = new ClassFinder(trees)
     (buffers, classFinder)
   }
